# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Component Repository API",
    "description": "REST API for managing CWL-based workflow components",
    "version": "1.0"
  },
  "paths": {
...
    "/workflows": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Create",
        "operationId": "create_workflows_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkflowRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid workflow file/archive, file too large, or unsupported inline step",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The workflow's or a configured tool's name already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/workflows/parse": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Parse Workflow",
        "operationId": "parse_workflow_workflows_parse_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ParseWorkflowRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseWorkflowResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid CWL/zip content or file too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/workflows/steps/{step_id}": {
      "patch": {
        "tags": [
          "workflows"
        ],
        "summary": "Update Step",
        "description": "Bind a step to a component - a tool, or another workflow to nest.",
        "operationId": "update_step_workflows_steps__step_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "step_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Step Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkflowStepRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStepDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not the creator of this workflow",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Step or component not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The workflow is published, the step runs an inline definition, the component is deprecated, or it would make the workflow contain itself",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workflows/steps/{step_id}/commands": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Execute Step Command",
        "operationId": "execute_step_command_workflows_steps__step_id__commands_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "step_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Step Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowStepCommandExecuteRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStepDto"
                }
              }
            }
          },
          "400": {
            "description": "Step has no matched component to confirm",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission for this command",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Step not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The workflow is published",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unknown command",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComponentConfigDto": {
        "properties": {
          "stepId": {
            "type": "string",
            "title": "Stepid"
          },
          "reuseComponentId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reusecomponentid"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "domains": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "items": {
              "enum": [
                "animal_behavior",
                "earth_observation",
                "other",
                "domain_agnostic"
              ]
            },
            "title": "Domains"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "formatLabels": {
            "items": {
              "$ref": "#/components/schemas/FormatLabelDto"
            },
            "type": "array",
            "title": "Formatlabels"
          }
        },
        "type": "object",
        "required": [
          "stepId"
        ],
        "title": "ComponentConfigDto",
        "description": "The user's decision for one previewed step, keyed by WorkflowStepPreviewDto.step_id.\n\nExactly one branch: reuse_component_id binds the step to an existing catalogue\ncomponent, or name+domain create a new tool from that step's CWL."
      },
      "ComponentCreatorDto": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "firstName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firstname"
          },
          "lastName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastname"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "firstName",
          "lastName"
        ],
        "title": "ComponentCreatorDto"
      },
...
      "ComponentSource": {
        "type": "string",
        "enum": [
          "automated_packaging",
          "manual_upload",
          "workflow_builder"
        ],
        "title": "ComponentSource"
      },
      "ComponentStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published",
          "deprecated"
        ],
        "title": "ComponentStatus",
        "description": "A component is staged privately until its creator publishes it. A published version\ncan later be deprecated: still readable, and everything already running it keeps\nworking, but it is no longer listed or offered for new use (see composite.lifecycle)."
      },
...
      "CreateWorkflowRequestDto": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "domains": {
            "items": {
              "enum": [
                "animal_behavior",
                "earth_observation",
                "other",
                "domain_agnostic"
              ]
            },
            "type": "array",
            "title": "Domains"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "componentConfigs": {
            "type": "string",
            "contentMediaType": "application/json",
            "contentSchema": {
              "items": {
                "$ref": "#/components/schemas/ComponentConfigDto"
              },
              "type": "array"
            },
            "title": "Componentconfigs",
            "default": "[]"
          }
        },
        "type": "object",
        "required": [
          "name",
          "domains",
          "file"
        ],
        "title": "CreateWorkflowRequestDto"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Link": {
        "properties": {
          "href": {
            "type": "string",
            "title": "Href"
          },
          "method": {
            "$ref": "#/components/schemas/HttpMethod",
            "default": "GET"
          }
        },
        "type": "object",
        "required": [
          "href"
        ],
        "title": "Link"
      },
...
      "ParseWorkflowRequestDto": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "ParseWorkflowRequestDto"
      },
      "ParseWorkflowResponseDto": {
        "properties": {
          "isZip": {
            "type": "boolean",
            "title": "Iszip"
          },
          "isSelfContained": {
            "type": "boolean",
            "title": "Isselfcontained"
          },
          "workflowName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflowname"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "stepCount": {
            "type": "integer",
            "title": "Stepcount"
          },
          "componentPreviews": {
            "items": {
              "$ref": "#/components/schemas/WorkflowStepPreviewDto"
            },
            "type": "array",
            "title": "Componentpreviews"
          },
          "externalRefs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Externalrefs"
          },
          "inlineOnlySteps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Inlineonlysteps"
          },
          "missingExternalRefs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missingexternalrefs"
          },
          "auxiliaryFiles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Auxiliaryfiles"
          },
          "missingImports": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missingimports"
          }
        },
        "type": "object",
        "required": [
          "isZip",
          "isSelfContained",
          "workflowName",
          "description",
          "stepCount",
          "componentPreviews",
          "externalRefs",
          "inlineOnlySteps",
          "missingExternalRefs",
          "auxiliaryFiles",
          "missingImports"
        ],
        "title": "ParseWorkflowResponseDto"
      },
...
      "StepMatchStatus": {
        "type": "string",
        "enum": [
          "suggested",
          "confirmed",
          "unmatched",
          "inline"
        ],
        "title": "StepMatchStatus"
      },
...
      "UpdateWorkflowStepRequestDto": {
        "properties": {
          "componentId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Componentid"
          }
        },
        "type": "object",
        "required": [
          "componentId"
        ],
        "title": "UpdateWorkflowStepRequestDto"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkflowDetailDto": {
        "properties": {
          "_links": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Link"
            },
            "type": "object",
            "title": "Links"
          },
          "kind": {
            "type": "string",
            "const": "workflow",
            "title": "Kind",
            "default": "workflow"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "authorName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authorname"
          },
          "createdBy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ComponentCreatorDto"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "repoUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repourl"
          },
          "repoCommitSha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repocommitsha"
          },
          "doi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doi"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "cwlContent": {
            "type": "string",
            "title": "Cwlcontent"
          },
          "ontologyUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ontologyurl"
          },
          "domains": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Domains"
          },
          "source": {
            "$ref": "#/components/schemas/ComponentSource"
          },
          "status": {
            "$ref": "#/components/schemas/ComponentStatus"
          },
          "deprecationNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deprecationnote"
          },
          "parameters": {
            "items": {
              "$ref": "#/components/schemas/ParameterDto"
            },
            "type": "array",
            "title": "Parameters"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "isFavorite": {
            "type": "boolean",
            "title": "Isfavorite"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/WorkflowStepDto"
            },
            "type": "array",
            "title": "Steps"
          },
          "draftId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draftid"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "authorName",
          "createdBy",
          "description",
          "repoUrl",
          "repoCommitSha",
          "doi",
          "version",
          "cwlContent",
          "ontologyUrl",
          "domains",
          "source",
          "status",
          "parameters",
          "createdAt",
          "updatedAt",
          "isFavorite",
          "steps",
          "draftId"
        ],
        "title": "WorkflowDetailDto"
      },
...
      "WorkflowStepCommandExecuteRequestDto": {
        "properties": {
          "command": {
            "$ref": "#/components/schemas/WorkflowStepCommandTypesApiV1",
            "description": "The specific action to perform on the step"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note",
            "description": "Optional note for the command execution"
          }
        },
        "type": "object",
        "required": [
          "command"
        ],
        "title": "WorkflowStepCommandExecuteRequestDto"
      },
...
      "WorkflowStepDto": {
        "properties": {
          "_links": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Link"
            },
            "type": "object",
            "title": "Links"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "stepId": {
            "type": "string",
            "title": "Stepid"
          },
          "runReference": {
            "type": "string",
            "title": "Runreference"
          },
          "stepOrder": {
            "type": "integer",
            "title": "Steporder"
          },
          "component": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ComponentSummaryDto"
              },
              {
                "type": "null"
              }
            ]
          },
          "matchStatus": {
            "$ref": "#/components/schemas/StepMatchStatus"
          },
          "matchScore": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matchscore"
          }
        },
        "type": "object",
        "required": [
          "id",
          "stepId",
          "runReference",
          "stepOrder",
          "component",
          "matchStatus",
          "matchScore"
        ],
        "title": "WorkflowStepDto"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}